requirements.txt
asgiref==3.8.1
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
Django==5.2.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
env==0.1.0
h11==0.16.0
Markdown==3.8
pycparser==2.22
PyMySQL==1.1.1
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.34.3
vnv==2.1
